Move build_href_url code into its own library
